// src/default/model/FilterOptionViewModel.ts
export interface FilterOptionViewModel {
  title: string;
  projectId: string;
  group: string;
  status: string;
  employee: string;
  layout?: number;
  _pageNumber?: number;
  _rowNumber?: number;
  projectType: string;
  fromDateBegin: string;
  fromDateEnd: string;
  toDateBegin: string;
  toDateEnd: string;
  dateCompeleteBegin: string;
  dateCompeleteEnd: string;
  isWorkRepeat: string;
  groupProject: string;
}